+++ test/test_simple_query.rb
@@ -44,16 +44,16 @@ module VespaRuby
       assert_equal "", yql
     end
 
-    test "ranking_profile is carried through build_request into post_body" do
+    test "rank_profile is carried through build_request into post_body" do
       q = SimpleQuery.query("meal break")
         .select("id")
         .type("weakAnd")
         .restrict("division")
         .default_index("ranktext")
-        .ranking_profile("bm25_anc_heavy")
+        .rank_profile("bm25_anc_heavy")
         .where(WhereOp.contains("jurisdiction", "CA"))
 
-      assert_equal "bm25_anc_heavy", q.ranking_profile_value
+      assert_equal "bm25_anc_heavy", q.rank_profile_value
 
       request = q.build_request(options: {hits: 10})
       # rubocop:disable Style/HashSyntax